Factor tree handler into own module
