+++ src/dbus_api/mod.rs
@@ -11,12 +11,12 @@ mod connection;
 mod consts;
 mod filesystem;
 mod pool;
+mod tree;
 mod types;
 mod udev;
 mod util;
 
 pub use self::{
-    connection::{DbusConnectionHandler, DbusTreeHandler},
-    udev::DbusUdevHandler,
+    connection::DbusConnectionHandler, tree::DbusTreeHandler, udev::DbusUdevHandler,
     util::create_dbus_handlers,
 };
